==> 1_data_ingestion.ipynb <==
import numpy as np
import pandas as pd
import requests, os
import time
import random
# import csv
# import datetime as dt
# import json
# import os
# import statistics

def toggle_vpn():
    global vpn_on
    if vpn_on: 
        os.system("pkill -x NordVPN\ IKE")
        print('Turning OFF')
    else:
        os.system("open /Applications/NordVPN\ IKE.app")
        print('Turning ON')
    time.sleep(10)
    vpn_on = not vpn_on
    print("Toggled VPN")

os.system("pkill -x openvpn")

# vpn_on = True
# os.system("open /Applications/NordVPN\ IKE.app")
from fake_useragent import UserAgent
ua = UserAgent()

user_agent = {'User-agent': ua.random}

def get_request(url, parameters=None, rate=60): 
    response = None
    while not response:
        try:
            response = requests.get(url=url, params=parameters, headers=user_agent)
            if not response:
                print('None response. Trying again in {} seconds...'.format(rate))
#                 toggle_vpn()
                time.sleep(rate)
        except:
            print('Request Exception. Trying again in 5 sec...')
            time.sleep(5)
            continue
        
    
    return response.json()

url = "https://steamspy.com/api.php"

for i in range(0,45):
    parameters = {"request": "all", "page": i}
    print("\'request\' = \'all\', 'page' = {}".format(i))
    time.sleep(60)
    json_data = get_request(url, parameters=parameters)
    app_list = pd.DataFrame.from_dict(json_data, orient='index')

    if not os.path.isfile('data/raw/app_list.csv'):
        app_list.to_csv('data/raw/app_list.csv', index=False)
    else: 
        app_list.to_csv('data/raw/app_list.csv', mode='a', header=False, index=False)



app_list.size

app_list = pd.read_csv('data/raw/app_list.csv', usecols=['appid'])['appid']
url = "https://steamspy.com/api.php"

for i in range(,app_list.size):
    parameters = {"request": "appdetails", "appid": app_list[i]}
    print(i)
    time.sleep(0.5)
    json_data = get_request(url, parameters=parameters, rate=1)
    try:
        json_data['tag_values'] = ', '.join(map(str, list(json_data['tags'].values())))
        json_data['tags'] = ', '.join(map(str, list(json_data['tags'].keys())))
    except:
        json_data['tags'] = ''
        json_data['tag_values'] = ''

        
    app_data = pd.DataFrame(json_data, index=[i])
    if not os.path.isfile('data/raw/app_data.csv'):
        app_data.to_csv('data/raw/app_data.csv', index=True)
    else: 
        app_data.to_csv('data/raw/app_data.csv', mode='a', header=False, index=True)
        

appid = app_list[44266]
parameters = {"appids": appid}
json_data = get_request(url, parameters=parameters, rate=2)
json_data

app_list = pd.read_csv('data/raw/app_list.csv', usecols=['appid'])['appid']
url = "http://store.steampowered.com/api/appdetails/"

for i in range(228,229):
    appid = app_list[i]
    parameters = {"appids": appid}
    print(i)
#     time.sleep(0.5)
    
    json_data = get_request(url, parameters=parameters, rate=10)[str(appid)]


    if json_data['success']:
        json_data = json_data['data']
    else:
        json_data = {'steam_appid': appid}
    
    all_keys = set(json_data.keys())
    desired_keys = ['type','name','steam_appid','required_age','is_free','controller_support','supported_languages','developers','publishers','platforms','metacritic','categories','genres','recommendations','release_date']
    
    missing_keys = set(desired_keys).difference(all_keys)
    for key in missing_keys:
        try:
            json_data[key] = None
        except:
            continue

    try:
        json_data['categories'] = ', '.join([x['description'] for x in json_data['categories']])
    except:
        json_data['categories'] = ''

    try:
        json_data['genres'] = ', '.join([x['description'] for x in json_data['genres']])
    except:
        json_data['genres'] = ''

    try:
        json_data['platforms'] = ', '.join([x[0] for x in json_data['platforms'].items() if x[1]])
    except:
        json_data['platforms'] = ''

    try:
        json_data['metacritic'] = json_data['metacritic']['score']
    except:
        json_data['metacritic'] = np.nan
    
    try:
        json_data['recommendations'] = json_data['recommendations']['total']
    except:
        json_data['recommendations'] = np.nan
        
    try:
        json_data['release_date'] = json_data['release_date']['date']
    except:
        try:
            if json_data['release_date']['coming_soon'] == True:
                json_data['release_date'] = 'coming soon'
            else:
                json_data['release_date'] = ''
        except:
            json_data['release_date'] = ''
    
    try:
        json_data['developers'] = ', '.join(json_data['developers'])
    except:
        json_data['developers'] = ''

    try:
        json_data['publishers'] = ', '.join(json_data['publishers'])
    except:
        json_data['publishers'] = ''

    json_data = {k: json_data[k] for k in desired_keys}    
    store_data = pd.DataFrame(json_data, index=[i])
    if not os.path.isfile('data/raw/store_data.csv'):
        store_data.to_csv('data/raw/store_data.csv', index=True)
    else:
        store_data.to_csv('data/raw/store_data.csv', mode='a', header=False, index=True)

app_list = pd.read_csv('data/raw/app_list.csv', usecols=['appid'])['appid']
url = "http://store.steampowered.com/api/appdetails/"

i=43735

appid = app_list[i]
parameters = {"appids": appid}
print(appid)
#     time.sleep(0.5)

json_data = get_request(url, parameters=parameters, rate=10)[str(appid)]


if json_data['success']:
    json_data = json_data['data']
else:
    json_data = {'steam_appid': appid}

all_keys = set(json_data.keys())
desired_keys = ['type','name','steam_appid','required_age','is_free','controller_support','supported_languages','developers','publishers','platforms','metacritic','categories','genres','recommendations','release_date']

missing_keys = set(desired_keys).difference(all_keys)
for key in missing_keys:
    try:
        json_data[key] = None
    except:
        continue

try:
    json_data['categories'] = ', '.join([x['description'] for x in json_data['categories']])
except:
    json_data['categories'] = ''

try:
    json_data['genres'] = ', '.join([x['description'] for x in json_data['genres']])
except:
    json_data['genres'] = ''

try:
    json_data['platforms'] = ', '.join([x[0] for x in json_data['platforms'].items() if x[1]])
except:
    json_data['platforms'] = ''

try:
    json_data['metacritic'] = json_data['metacritic']['score']
except:
    json_data['metacritic'] = np.nan

try:
    json_data['recommendations'] = json_data['recommendations']['total']
except:
    json_data['recommendations'] = np.nan

try:
    json_data['release_date'] = json_data['release_date']['date']
except:
    try:
        if json_data['release_date']['coming_soon'] == True:
            json_data['release_date'] = 'coming soon'
        else:
            json_data['release_date'] = ''
    except:
        json_data['release_date'] = ''

try:
    json_data['developers'] = ', '.join(json_data['developers'])
except:
    json_data['developers'] = ''

try:
    json_data['publishers'] = ', '.join(json_data['publishers'])
except:
    json_data['publishers'] = ''

json_data = {k: json_data[k] for k in desired_keys}    
#     print(json_data)
store_data = pd.DataFrame(json_data, index=[i])
if not os.path.isfile('data/raw/store_data_english_fixes.csv'):
    store_data.to_csv('data/raw/store_data_english_fixes.csv', index=True)
else:
    store_data.to_csv('data/raw/store_data_english_fixes.csv', mode='a', header=False, index=True)
