==> 2_data_cleaning.ipynb <==
import numpy as np
import pandas as pd
import requests, os
import time
import random
import seaborn as sns
# import csv
import datetime as dt
import regex as re
# import json
# import os
# import statistics
pd.options.display.max_columns = 30
pd.options.display.max_rows = 200

"".split(', ')

store_df = pd.read_csv('data/raw/store_data.csv',index_col=0)
english_fixes_df = pd.read_csv('data/raw/store_data_english_fixes.csv',index_col=0)
store_df.update(english_fixes_df)

store_df = store_df[~store_df['name'].isnull()]
store_df = store_df[store_df['type']=='game']
store_df = store_df.drop(columns=['type'])
store_df.dropna(subset=['supported_languages'], inplace=True)
store_df['required_age'] = store_df['required_age'].astype(str)
store_df['required_age'] = store_df['required_age'].str.replace('\.0','').str.replace('\+','').str.replace('1818','18').astype(int)
store_df = store_df[~(store_df['supported_languages'].str.contains('(\\r\\n|\#)',regex=True))]
store_df = store_df.drop(index=[601,12472])



store_df['supported_languages'] = (store_df['supported_languages']
                                   .str.replace(',',', ')
                                   .str.replace(',  ',', ')
                                   .str.replace(' \(full audio\)','*')
                                   .str.replace(' \(text only\)','')
                                   .str.replace('\;','')
                                   .str.replace('  ','')
                                   .str.replace('\<br\>\<strong\>\*\<\/strong\>languages with full audio support','')
                                   .str.replace('\<strong\>\*\<\/strong\>','*')
                                   .str.replace(',,',',')
                                  )



store_df['supported_languages_audio'] = store_df['supported_languages'].str.split(', ')
store_df['supported_languages'] = store_df['supported_languages'].str.replace('*','')
store_df['release_date'] = pd.to_datetime(store_df['release_date'],errors='coerce')
store_df.dropna(subset=['release_date'], inplace=True)
store_df['age_days'] = (dt.datetime.now() - store_df['release_date']).dt.days
store_df['publishers'] = (store_df['publishers']
                          .str.replace(', *\.*(i\.*n\.*c|l\.*t\.*d|l\.*l\.*c|s\.r\.o|s\.c|l\.*d\.*a|co\.)\.*', '', flags=re.IGNORECASE, regex=True)
                          .str.replace(', *\.*(ACE|SL|a\.s)\.*', '', regex=True)
                          .str.replace(',[^,]*\((mac|linux)( *,*/* *(mac|linux))*\)','', flags=re.IGNORECASE, regex=True)
                          .str.replace(',,',',', flags=re.IGNORECASE, regex=True)
                          .str.replace(', *$', '', flags=re.IGNORECASE, regex=True)
                         )
store_df['developers'] = (store_df['developers']
                          .str.replace(', *\.*(i\.*n\.*c|l\.*t\.*d|l\.*l\.*c|s\.r\.o|s\.c|l\.*d\.*a|co\.)\.*', '', flags=re.IGNORECASE, regex=True)
                          .str.replace(', *\.*(ACE|SL|a\.s)\.*', '', regex=True)
                          .str.replace(',[^,]*\((mac|linux)( *,*/* *(mac|linux))*\)','', flags=re.IGNORECASE, regex=True)
                          .str.replace(',,',',', flags=re.IGNORECASE, regex=True)
                          .str.replace(', *$', '', flags=re.IGNORECASE, regex=True)
                         )

store_df = store_df.dropna(subset=['developers','categories','genres'])

def strong_only(x):
    return [y[:-1] for y in x if y[-1]=='*']

store_df['supported_languages_audio'] = store_df['supported_languages_audio'].apply(strong_only)
store_df['supported_languages_audio'] = store_df['supported_languages_audio'].apply(lambda x: ', '.join(x))


store_df['controller_support'] = store_df['controller_support'].apply(lambda x: 1 if x=='full' else 0).astype(int)
store_df['is_free'] = store_df['is_free'].apply(lambda x: 1 if x==True else 0).astype(int)
print(len(store_df))

# len(
# store_df['developers'][store_df['developers'].str.contains('$', flags=re.IGNORECASE, regex=True)].head(100)


store_df[(store_df['supported_languages'].str.contains('DutchEnglish',regex=True))]

# )

spy_df = pd.read_csv('data/raw/app_data.csv', index_col=0)
spy_df['owners'] = spy_df['owners'].str.replace('\ ','').str.replace('\,','').str.split('\.\.')
spy_df['owners'] = spy_df['owners'].apply(lambda x: x[1])
spy_df['owners'] = spy_df['owners'].astype(int)
spy_df

combined = pd.merge(spy_df.drop(columns=['name','genre','developer','publisher','score_rank','languages']),store_df,right_on='steam_appid',left_on='appid').set_index('appid')

combined = combined[['name', 'developers', 'publishers', 'platforms', 'categories', 'genres', 'tags', 'tag_values',  
 'positive', 'negative', 'userscore', 'owners',
       'average_forever', 'average_2weeks', 'median_forever', 'median_2weeks', 'ccu', 
       'is_free', 'price', 'initialprice', 'discount',
       'required_age', 'controller_support', 'supported_languages', 'supported_languages_audio',
       'metacritic', 'recommendations',
       'release_date','age_days']]

combined.to_csv('data/combined.csv')
