==> 4_base_model.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns

from sklearn.preprocessing import MultiLabelBinarizer
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.pipeline import Pipeline
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.metrics import f1_score, recall_score, precision_score, fbeta_score

from sklearn.model_selection import train_test_split

from sklearn.base import TransformerMixin #gives fit_transform method for free

class MultiLabelEncoder(TransformerMixin):
    def __init__(self, *args, **kwargs):
        self.encoder = MultiLabelBinarizer(*args, **kwargs)
    def fit(self, x, y=0):
        self.encoder.fit(x)
        self.classes_ = self.encoder.classes_
        return self
    def transform(self, x, y=0):
        return self.encoder.transform(x)

df = pd.read_csv('data/combined_cleaned.csv')
df['genres'] = df['genres'].str.split(', ')
df['tags'] = df['tags'].str.split(', ')
df['categories'] = df['categories'].str.split(', ')



df['target'] = (df['average_2weeks'] > 500).astype(int)

df['target'].describe()

mlb.fit_transform(df['genres'])


rf =  RandomForestClassifier(1000,class_weight={1: 10, 0: 1})
rf.fit(X_train,y_train)
y_predict = rf.predict(X_test)
print('precision: ', precision_score(y_test,y_predict))
print('recall: ', recall_score(y_test,y_predict))
print('f1 :', f1_score(y_test, y_predict))

mlb = MultiLabelEncoder()

rf =  RandomForestClassifier(1000,class_weight={1: 20, 0: 1})
rf.fit(X_train,y_train)
y_predict = rf.predict(X_test)
print('precision: ', precision_score(y_test,y_predict))
print('recall: ', recall_score(y_test,y_predict))
print('f1 :', f1_score(y_test, y_predict))

mlb = MultiLabelEncoder()

rf =  RandomForestClassifier(1000,class_weight='balanced')
rf.fit(X_train,y_train)
y_predict = rf.predict(X_test)
print('precision: ', precision_score(y_test,y_predict))
print('recall: ', recall_score(y_test,y_predict))
print('f1 :', f1_score(y_test, y_predict))

mlb = MultiLabelEncoder()

rf =  RandomForestClassifier(1000)
rf.fit(X_train,y_train)
y_predict = rf.predict(X_test)
print('precision: ', precision_score(y_test,y_predict))
print('recall: ', recall_score(y_test,y_predict))
print('f1 :', f1_score(y_test, y_predict))

y_test.describe()
